refactor(api): move tenants, virtual_keys, routing, providers_by_project endpoints
Commit 5 of the dashboard fold-in. Lifts 7 endpoints out of
dashboard/api/main.py into four new files under
server/api/dashboard/:

- providers_by_project.py: GET /api/providers/by-project
                           (carries its own _LOCAL_PROVIDER_NAMES copy
                            since dashboard/status.py already owns one
                            and the two are not coupled enough to share
                            a third module).
- tenants.py:              GET /api/tenants, /api/tenants/{id}
- virtual_keys.py:         GET /api/virtual_keys, POST same,
                           POST /api/virtual_keys/{id}/revoke
- routing.py:              GET /api/routing/observations

Behaviour preserved verbatim. virtual_keys / tenants / routing keep
their _ensure_initialized + _conn.session() storage-facade pattern;
providers_by_project keeps the YAML-then-DB precedence loop and the
Fernet decrypt+mask path.

test_routing_endpoint.py drops its _FakeGateway shim in favour of a
real Gateway built from temp_config (the daemon's build_app needs
the layered DI stack that the fake didn't model). test_providers_by_project.py
swaps both fixtures to use build_app. The one inline test in
test_dashboard_api.py for /api/virtual_keys already runs through the
build_app-backed client fixture from commit 2.

ruff check src + uv run mypy + uv run pytest -q all green:
1666 passed, 4 skipped.
