+++ src/voicegateway/server/routes.py
@@ -51,13 +51,25 @@ from voicegateway.server.api.dashboard import (
     projects as dashboard_projects,
 )
 from voicegateway.server.api.dashboard import (
+    providers_by_project as dashboard_providers_by_project,
+)
+from voicegateway.server.api.dashboard import (
     replay as dashboard_replay,
+)
+from voicegateway.server.api.dashboard import (
+    routing as dashboard_routing,
 )
 from voicegateway.server.api.dashboard import (
     sessions as dashboard_sessions,
 )
 from voicegateway.server.api.dashboard import (
     status as dashboard_status,
+)
+from voicegateway.server.api.dashboard import (
+    tenants as dashboard_tenants,
+)
+from voicegateway.server.api.dashboard import (
+    virtual_keys as dashboard_virtual_keys,
 )
 
 system_router = APIRouter()
@@ -86,6 +98,10 @@ dashboard_router.include_router(dashboard_sessions.router)
 dashboard_router.include_router(dashboard_metrics.router)
 dashboard_router.include_router(dashboard_guardrails.router)
 dashboard_router.include_router(dashboard_replay.router)
+dashboard_router.include_router(dashboard_providers_by_project.router)
+dashboard_router.include_router(dashboard_tenants.router)
+dashboard_router.include_router(dashboard_virtual_keys.router)
+dashboard_router.include_router(dashboard_routing.router)
 
 
 __all__ = ["api_router", "dashboard_router", "system_router"]
